% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ammatit_1270.R
\docType{data}
\name{data_ammatit_1270}
\alias{data_ammatit_1270}
\title{Avoimet työpaikat, työttömät ja työvoima ammateittain koko maassa}
\format{
A data frame
\describe{
\item{Avoimet_tyopaikat}{}
\item{Ammattiryhma}{}
\item{Tyottomat}{}
\item{time}{}
}
}
\source{
\url{http://pxnet2.stat.fi/PXWeb/api/v1/fi/StatFin/tym/tyonv/kk/statfin_tyonv_pxt_1270.px}
}
\usage{
data_ammatit_1270
}
\description{
Avoimet työpaikat, työttömät ja työvoima ammateittain koko maassa
}
\keyword{datasets}
