% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_kokomaa_2310.R
\docType{data}
\name{data_kokomaa_2310}
\alias{data_kokomaa_2310}
\title{Tietoja avoimista työpaikoista koko mmaa}
\format{
A data frame
\describe{
\item{Avoimet_tyopaikat_kauden_aikana}{}
\item{Uudet_avoimet_tyopaikat_kauden_aikana}{}
\item{Tayttyneet_tyopaikat_kauden_aikana}{}
\item{Tayttyneiden_tyopaikkojen_avoinnaolon_km_kesto_paivaa}{}
\item{time}{}
\item{seutukunta}{}
}
}
\source{
\url{http://pxnet2.stat.fi/PXWeb/api/v1/fi/StatFin/tym/tyonv/kk/statfin_tyonv_pxt_2310.px}
}
\usage{
data_kokomaa_2310
}
\description{
Tietoja avoimista työpaikoista koko mmaa
}
\keyword{datasets}
