% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_seutukunta_1001.R
\docType{data}
\name{data_seutukunta_1001}
\alias{data_seutukunta_1001}
\title{Avoimet työpaikat, työttömät ja työvoima seutukunnittain}
\format{
A data frame
\describe{
\item{Avoimet_tyopaikat}{}
\item{Tyovoima}{}
\item{Tyottomat}{}
\item{time}{}
\item{seutukunta}{}
}
}
\source{
\url{http://pxnet2.stat.fi/PXWeb/api/v1/fi/StatFin/tym/tyonv/kk/statfin_tyonv_pxt_1001.px}
}
\usage{
data_seutukunta_1001
}
\description{
Avoimet työpaikat, työttömät ja työvoima seutukunnittain
}
\keyword{datasets}
