% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_region.R
\name{filter_region}
\alias{filter_region}
\title{Filters the desired region from the dataset}
\usage{
filter_region(data, region)
}
\arguments{
\item{data}{data.frame, contains the regional data}

\item{region}{character, the desired region.}
}
\value{
data.frame
}
\description{
Filters the desired region from the dataset
}
