% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_region_level.R
\name{filter_region_level}
\alias{filter_region_level}
\title{Filters the desired region from the dataset}
\usage{
filter_region_level(data, region)
}
\arguments{
\item{data}{data.frame, contains the regional data}

\item{region}{character, the desired region.}
}
\value{
data.frame
}
\description{
#' Tilastokeskuksen taulukoissa on usein tietoja eri aggregointitasolla. Koska suurimmalla
aggregointitasolla on vähiten suojattuja tietoja (vähemmän kuin 5 havaintoa) tulisi anlyysissä
käyttää aina suurinta mahdollista aggregointitasoa siten että, jos haluaa aggregoida alueita,
joita pystyy aggregoimaan seutukunnista, ei tulisi aggregoida kuntatiedoista, vaan seutukunta
tiedoista. Jos aggregointiin voi käyttää maakuntia, vielä parempi.
}
\details{
Kunnille ei tule koodeja suoraan pxwebin aineistosta
}
