% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_beveridge_curve.R
\name{plot_beveridge_curve}
\alias{plot_beveridge_curve}
\title{Plots a Beveridge curve}
\usage{
plot_beveridge_curve(
  df,
  series_type = "trend",
  number_type = "relative",
  colors = c("#00627D", "#00713D")
)
}
\arguments{
\item{colors}{}
}
\value{

}
\description{
Plots a Beveridge curve
}
